Modifying pom.xml for 'clean start', deleting useless files and updating gitignore
It's been nearly a year since the last commit on this repo. I'll try to keep working on it in my spare time. I tried to run the code as it was at the time of the last commit, but the IDE returned a few errors related to dependencies, so I've modified the pom.xml in order to establish the mininum that I need in order to successfully start the service. From now on, I'll merge this branch in main, and start developing from another branch.