Adjusting BookControler endpoint implementation, Refactoring BookQuery and AdviceHandler, Adding default entries for in-memory DB and organazing pom.xml
BookController: Should throw ResponseStatusException instead of NoResourceFoundException, because both exceptions are treated almost equally, with the single difference being that, when a NoResourceFoundException is thrown, the Handler will return a message saying that the URL for the request might be wrong, while in other cases (such as a non-existing book), the message willsay that no book was found. [Even though the 404 status code can be used for both cases, I'll probably handle the first situation as a bad request, returning 400 instead of 404].

BookQuery: Refactoring switch case to be more concise

ControllerAdviceHandler: For now, refactoring a lambda.

Application: Using commandLineRunner for creating entities on startup, in order to test endpoints

pom.xml : Adjusting dependencies version and moving to separate those that are currently being used

SecurityConfig: Adding ADMIN role for testing deletions